Renamed project to lxcri.
