+++ cgroup_test.go
@@ -1,4 +1,4 @@
-package lxcontainer
+package lxcri
 
 import (
 	"testing"
